@@ -1,5 +1,5 @@
-<!--
-Copyright 2021 Coveo
+﻿<!--
+Copyright 2024 Coveo
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -17,27 +17,27 @@ limitations under the License.
 <Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Library</OutputType>
-    <TargetFrameworks>net461;netcoreapp2.1</TargetFrameworks>
     <PackageId>sfdx4csharp</PackageId>
-    <PackageVersion>0.9.0</PackageVersion>
-    <Authors>Marc-Antoine Veilleux;Denis Shelgunov;Benjamin Ross;Felix Pelletier</Authors>
-    <Owners>Marc-Antoine Veilleux;Denis Shelgunov;Felix Pelletier</Owners>
+    <PackageVersion>1.0.0</PackageVersion>
+    <Authors>Denis Shelgunov;Vincent Gregoire;Rozerin Akkus</Authors>
+    <Owners>Denis Shelgunov;Vincent Gregoire;Rozerin Akkus</Owners>
     <Title>SFDX for C#</Title>
-    <Description>A simple wrapper in C# for SFDX</Description>
+    <Description>A simple wrapper in C# for the offcial Salesfoce Cli</Description>
     <ProjectUrl>https://github.com/coveo/sfdx4csharp</ProjectUrl>
     <LicenseUrl>https://github.com/coveo/sfdx4csharp/LICENSE.md</LicenseUrl>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
-    <Copyright>Copyright 2021 Coveo</Copyright>
-    <PackageTags>SFDX;C#;Salesforce DX;Wrapper;Library;Coveo</PackageTags>
+    <Copyright>Copyright 2024 Coveo</Copyright>
+    <PackageTags>SFDX;SF;Cli;C#;Salesforce DX;Wrapper;Library;Coveo</PackageTags>
   </PropertyGroup>
   
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFrameworks>net461;netcoreapp2.1</TargetFrameworks>
+    <TargetFrameworks>net6.0</TargetFrameworks>
     <NoWin32Manifest>true</NoWin32Manifest>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
   </PropertyGroup>
 
   <ItemGroup>
-      <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+      <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 </Project>
