@@ -1,5 +1,4 @@
-// Generated on Thu Aug 05 2021 using sfdx-cli/7.110.0 win32-x64 node-v16.2.0. DO NOT MODIFY
-// Copyright (c) 2005-2021, Coveo Solutions Inc.
+// Generated on Wed May 29 2024 using @salesforce/cli/2.42.6 win32-x64 node-v20.12.2. DO NOT MODIFY
 
 using sfdx4csharp.Client.Core;
 using sfdx4csharpClient.Core;
@@ -8,79 +7,68 @@ using sfdx4csharpClient.Core.Attributes;
 namespace sfdx4csharpClient
 {
     /// <summary>
-    /// Options for the method templateCreate of class Analytics.
+    /// Options for the method GenerateTemplate of class Analytics.
     /// </summary>
-    public class AnalyticsTemplateCreateOptions : SFDXOptions
+    public class AnalyticsGenerateTemplateOptions : SfdxOptions
     {
         /// <summary>
         /// [Required] Raw string parameters for the command. EX: 'name=value' expressions or parameters without flags.
         /// </summary>
         [SwitchName("")]
-        public string expression { get; set; }
+        public string Expression { get; set; }
 
         /// <summary>
-        /// [Required] The name of the Analytics template.
+        /// [Required] No description for name
         /// </summary>
-        [SwitchName("--templatename")]
-        public string templatename { get; set; }
+        [SwitchName("--name")]
+        public string Name { get; set; }
 
         /// <summary>
-        /// [Optional] Logging level for this command invocation
+        /// [Required] The location can be an absolute path or relative to the current working directory. The default is the current directory.
         /// </summary>
-        [SwitchName("--loglevel")]
-        public LogLevel? loglevel { get; set; }
-
-        /// <summary>
-        /// [Required] The directory to store the newly created files. The location can be an absolute path or relative to the current working directory. The default is the current directory.
-        /// </summary>
-        [SwitchName("--outputdir")]
-        public string outputdir { get; set; }
+        [SwitchName("--output-dir")]
+        public string OutputDir { get; set; }
 
         /// <summary>
         /// [Required] Override the api version used for api requests made by this command
         /// </summary>
-        [SwitchName("--apiversion")]
-        public string apiversion { get; set; }
+        [SwitchName("--api-version")]
+        public string ApiVersion { get; set; }
+
+        /// <summary>
+        /// [Required] No description for loglevel
+        /// </summary>
+        [SwitchName("--loglevel")]
+        public string Loglevel { get; set; }
     }
 
     /// <summary>
     /// Analytics
     /// </summary>
-    [CommandClass("force:analytics")]
-    public class Analytics : SFDXCommand
+    [CommandClass("analytics")]
+    public class Analytics : SfdxCommand
     {
         /// Constructor.
-        public Analytics(CommandExecutioner p_CommandExecutioner)
-                : base(p_CommandExecutioner)
+        public Analytics(CommandExecutioner commandExecutioner)
+                : base(commandExecutioner)
         {
         }
 
         /// <summary>
-        /// add an Analytics template to your workspace
-        /// If not supplied, the apiversion, template, and outputdir use default values.
-        /// The outputdir can be an absolute path or relative to the current working directory.
-        /// 
+        /// The metadata files associated with the Analytics template must be contained in a parent directory called "waveTemplates" in your package directory. Either run this command from an existing directory of this name, or use the --output-dir flag to generate one or point to an existing one.
         /// </summary>
         /// <remarks>
-        /// add an Analytics template to your workspace
-        /// If not supplied, the apiversion, template, and outputdir use default values.
-        /// The outputdir can be an absolute path or relative to the current working directory.
-        /// 
+        /// The metadata files associated with the Analytics template must be contained in a parent directory called "waveTemplates" in your package directory. Either run this command from an existing directory of this name, or use the --output-dir flag to generate one or point to an existing one.
         /// </remarks>
         /// <example>
-        /// If not supplied, the apiversion, template, and outputdir use default values.
-        /// The outputdir can be an absolute path or relative to the current working directory.
         /// 
-        /// Examples:
-        ///    $ sfdx force:analytics:template:create -n myTemplate -d outputdir
-        /// 
-        /// $ sfdx force:analytics:template:create -n myTemplate -d outputdir
-        /// force:analytics:template:create -n <string> [-d <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]
+        /// Generate the metadata files for a simple Analytics template file called myTemplate in the force-app/main/default/waveTemplates directory:
+        /// <%= config.bin %> <%= command.id %> --name myTemplate --output-dir force-app/main/default/waveTemplates
         /// </example>
-        [Command("template:create")]
-        public SFDXResponse TemplateCreate(AnalyticsTemplateCreateOptions p_Options)
+        [Command("generate template")]
+        public SfdxResponse GenerateTemplate(AnalyticsGenerateTemplateOptions options)
         {
-            return ExecuteCommand<AnalyticsTemplateCreateOptions>(nameof(TemplateCreate), p_Options);
+            return ExecuteCommand(nameof(GenerateTemplate), options);
         }
     }
 }
