@@ -1,5 +1,4 @@
-// Generated on Thu Aug 05 2021 using sfdx-cli/7.110.0 win32-x64 node-v16.2.0. DO NOT MODIFY
-// Copyright (c) 2005-2021, Coveo Solutions Inc.
+// Generated on Wed May 29 2024 using @salesforce/cli/2.42.6 win32-x64 node-v20.12.2. DO NOT MODIFY
 
 using sfdx4csharp.Client.Core;
 using sfdx4csharpClient.Core;
@@ -8,277 +7,266 @@ using sfdx4csharpClient.Core.Attributes;
 namespace sfdx4csharpClient
 {
     /// <summary>
-    /// Options for the method create of class Community.
+    /// Options for the method Create of class Community.
     /// </summary>
-    public class CommunityCreateOptions : SFDXOptions
+    public class CommunityCreateOptions : SfdxOptions
     {
         /// <summary>
         /// [Required] Raw string parameters for the command. EX: 'name=value' expressions or parameters without flags.
         /// </summary>
         [SwitchName("")]
-        public string expression { get; set; }
+        public string Expression { get; set; }
 
         /// <summary>
-        /// [Required] The URL to append to the domain that you created when you enabled Digital Experiences for this org. For example, if your domain name is UniversalTelco.force.com and you’re creating a customer site, enter "customers" to create the unique URL 
-        /// UniversalTelco.force.com/customers.
+        /// [Required] No description for target-org
         /// </summary>
-        [SwitchName("--urlpathprefix")]
-        public string urlpathprefix { get; set; }
+        [SwitchName("--target-org")]
+        public string TargetOrg { get; set; }
 
         /// <summary>
-        /// [Required] The template to use to create the site, such as the Customer Service template. Run force:community:template:list to see which templates are available in your org.
+        /// [Required] An example of a template is Customer Service. Run the "community template list" command to see which templates are available in your org.
         /// </summary>
-        [SwitchName("--templatename")]
-        public string templatename { get; set; }
+        [SwitchName("--template-name")]
+        public string TemplateName { get; set; }
 
         /// <summary>
-        /// [Required] The name of the site to create.
+        /// [Required] No description for name
         /// </summary>
         [SwitchName("--name")]
-        public string name { get; set; }
+        public string Name { get; set; }
 
         /// <summary>
-        /// [Optional] Logging level for this command invocation
+        /// [Optional] Format output as json.
         /// </summary>
-        [SwitchName("--loglevel")]
-        public LogLevel? loglevel { get; set; }
+        [SwitchName("--json")]
+        public bool? Json { get; set; }
 
         /// <summary>
-        /// [Required] Username or alias for the target org; overrides default target org
+        /// [Required] No description for flags-dir
         /// </summary>
-        [SwitchName("--targetusername")]
-        public string targetusername { get; set; }
+        [SwitchName("--flags-dir")]
+        public string FlagsDir { get; set; }
 
         /// <summary>
-        /// [Required] Override the api version used for api requests made by this command
+        /// [Required] For example, if your domain name is https://MyDomainName.my.site.com and you create a customer site, enter 'customers' to create the unique URL https://MyDomainName.my.site.com/customers.
         /// </summary>
-        [SwitchName("--apiversion")]
-        public string apiversion { get; set; }
+        [SwitchName("--url-path-prefix")]
+        public string UrlPathPrefix { get; set; }
 
         /// <summary>
-        /// [Required] The description of the site. The description displays in Digital Experiences - All Sites in Setup and helps with site identification.
+        /// [Required] The description displays in Digital Experiences - All Sites in Setup and helps with site identification.
         /// </summary>
         [SwitchName("--description")]
-        public string description { get; set; }
+        public string Description { get; set; }
+
+        /// <summary>
+        /// [Optional] No description for loglevel
+        /// </summary>
+        [SwitchName("--loglevel")]
+        public LogLevel? Loglevel { get; set; }
+
+        /// <summary>
+        /// [Required] Override the api version used for api requests made by this command
+        /// </summary>
+        [SwitchName("--api-version")]
+        public string ApiVersion { get; set; }
     }
 
     /// <summary>
-    /// Options for the method publish of class Community.
+    /// Options for the method ListTemplate of class Community.
     /// </summary>
-    public class CommunityPublishOptions : SFDXOptions
+    public class CommunityListTemplateOptions : SfdxOptions
     {
         /// <summary>
         /// [Required] Raw string parameters for the command. EX: 'name=value' expressions or parameters without flags.
         /// </summary>
         [SwitchName("")]
-        public string expression { get; set; }
+        public string Expression { get; set; }
 
         /// <summary>
-        /// [Required] The name of the Experience Builder site that you want to publish.
+        /// [Required] No description for target-org
+        /// </summary>
+        [SwitchName("--target-org")]
+        public string TargetOrg { get; set; }
+
+        /// <summary>
+        /// [Optional] Format output as json.
+        /// </summary>
+        [SwitchName("--json")]
+        public bool? Json { get; set; }
+
+        /// <summary>
+        /// [Required] No description for flags-dir
+        /// </summary>
+        [SwitchName("--flags-dir")]
+        public string FlagsDir { get; set; }
+
+        /// <summary>
+        /// [Required] Override the api version used for api requests made by this command
+        /// </summary>
+        [SwitchName("--api-version")]
+        public string ApiVersion { get; set; }
+
+        /// <summary>
+        /// [Optional] No description for loglevel
+        /// </summary>
+        [SwitchName("--loglevel")]
+        public LogLevel? Loglevel { get; set; }
+    }
+
+    /// <summary>
+    /// Options for the method Publish of class Community.
+    /// </summary>
+    public class CommunityPublishOptions : SfdxOptions
+    {
+        /// <summary>
+        /// [Required] Raw string parameters for the command. EX: 'name=value' expressions or parameters without flags.
+        /// </summary>
+        [SwitchName("")]
+        public string Expression { get; set; }
+
+        /// <summary>
+        /// [Required] No description for target-org
+        /// </summary>
+        [SwitchName("--target-org")]
+        public string TargetOrg { get; set; }
+
+        /// <summary>
+        /// [Required] No description for name
         /// </summary>
         [SwitchName("--name")]
-        public string name { get; set; }
+        public string Name { get; set; }
 
         /// <summary>
-        /// [Optional] Logging level for this command invocation
+        /// [Optional] Format output as json.
         /// </summary>
-        [SwitchName("--loglevel")]
-        public LogLevel? loglevel { get; set; }
+        [SwitchName("--json")]
+        public bool? Json { get; set; }
 
         /// <summary>
-        /// [Required] Username or alias for the target org; overrides default target org
+        /// [Required] No description for flags-dir
         /// </summary>
-        [SwitchName("--targetusername")]
-        public string targetusername { get; set; }
-
-        /// <summary>
-        /// [Required] Override the api version used for api requests made by this command
-        /// </summary>
-        [SwitchName("--apiversion")]
-        public string apiversion { get; set; }
-    }
-
-    /// <summary>
-    /// Options for the method templateList of class Community.
-    /// </summary>
-    public class CommunityTemplateListOptions : SFDXOptions
-    {
-        /// <summary>
-        /// [Required] Raw string parameters for the command. EX: 'name=value' expressions or parameters without flags.
-        /// </summary>
-        [SwitchName("")]
-        public string expression { get; set; }
-
-        /// <summary>
-        /// [Optional] Logging level for this command invocation
-        /// </summary>
-        [SwitchName("--loglevel")]
-        public LogLevel? loglevel { get; set; }
-
-        /// <summary>
-        /// [Required] Username or alias for the target org; overrides default target org
-        /// </summary>
-        [SwitchName("--targetusername")]
-        public string targetusername { get; set; }
+        [SwitchName("--flags-dir")]
+        public string FlagsDir { get; set; }
 
         /// <summary>
         /// [Required] Override the api version used for api requests made by this command
         /// </summary>
-        [SwitchName("--apiversion")]
-        public string apiversion { get; set; }
+        [SwitchName("--api-version")]
+        public string ApiVersion { get; set; }
+
+        /// <summary>
+        /// [Optional] No description for loglevel
+        /// </summary>
+        [SwitchName("--loglevel")]
+        public LogLevel? Loglevel { get; set; }
     }
 
     /// <summary>
     /// Community
     /// </summary>
-    [CommandClass("force:community")]
-    public class Community : SFDXCommand
+    [CommandClass("community")]
+    public class Community : SfdxCommand
     {
         /// Constructor.
-        public Community(CommandExecutioner p_CommandExecutioner)
-                : base(p_CommandExecutioner)
+        public Community(CommandExecutioner commandExecutioner)
+                : base(commandExecutioner)
         {
         }
 
         /// <summary>
-        /// creates an Experience Cloud site using a template
+        /// Run the "community list template" command to see the templates available in your org. See 'Which Experience Cloud Template Should I Use?' in Salesforce Help for more information about the different template types available. (https://help.salesforce.com/s/articleView?id=sf.siteforce_commtemp_intro.htm&type=5)
         /// 
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
+        /// When you create a site with the Build Your Own (LWR) template, you must also specify the AuthenticationType value using the format templateParams.AuthenticationType=value, where value is AUTHENTICATED or AUTHENTICATED_WITH_PUBLIC_ACCESS_ENABLED. Name and values are case-sensitive. See 'DigitalExperienceBundle' in the Metadata API Developer Guide for more information. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_digitalexperiencebundle.htm)
         /// 
-        /// When creating a site with the Build Your Own (LWR) template, you must also specify the AuthenticationType value using the format templateParams.AuthenticationType=value, where value is AUTHENTICATED, UNAUTHENTICATED, or AUTHENTICATED_WITH_PUBLIC_ACCESS. Name and values are case-sensitive. See "ExperienceBundle" in the Metadata API Developer Guide for more information.
+        /// The site creation process is an async job that generates a jobId. To check the site creation status, query the BackgroundOperation object and enter the jobId as the Id. See ‘BackgroundOperation’ in the Object Reference for the Salesforce Platform for more information. (https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_backgroundoperation.htm)
         /// 
-        /// When you execute this command, it creates the site in preview status, which means that it isn’t yet live. After you finish building your site, you can make it live.
+        /// If the job doesn’t complete within 10 minutes, it times out. You receive an error message and must restart the site creation process. Completed jobs expire after 24 hours and are removed from the database.
         /// 
-        /// If you have an Experience Builder site, publish the site using the sfdx force:community:publish command to make it live.
+        /// When you run this command, it creates the site in preview status, which means that the site isn't yet live. After you finish building your site, you can make it live.
         /// 
-        /// If you have a Salesforce Tabs + Visualforce site, activate the site to make it live by updating the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
+        /// If you have an Experience Builder site, publish the site using the "community publish" command to make it live.
         /// 
-        /// For Experience Builder sites, activating the site just sends out a welcome email to site members.
+        /// If you have a Salesforce Tabs + Visualforce site, to activate the site and make it live, update the status field of the Network type in Metadata API. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_network.htm) Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
         /// 
-        /// Examples:
-        ///    $ sfdx force:community:create --name "My Customer Site" --templatename "Customer Service" --urlpathprefix customers --description "My customer site"
-        ///    $ sfdx force:community:create -n partnercentral -t "Partner Central" -p partners
-        ///    $ sfdx force:community:create -n lwrsite -t "Build Your Own (LWR)" -p lwrsite templateParams.AuthenticatedType=UNAUTHENTICATED
+        /// For Experience Builder sites, activating the site sends a welcome email to site members.
         /// </summary>
         /// <remarks>
-        /// creates an Experience Cloud site using a template
+        /// Run the "community list template" command to see the templates available in your org. See 'Which Experience Cloud Template Should I Use?' in Salesforce Help for more information about the different template types available. (https://help.salesforce.com/s/articleView?id=sf.siteforce_commtemp_intro.htm&type=5)
         /// 
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
+        /// When you create a site with the Build Your Own (LWR) template, you must also specify the AuthenticationType value using the format templateParams.AuthenticationType=value, where value is AUTHENTICATED or AUTHENTICATED_WITH_PUBLIC_ACCESS_ENABLED. Name and values are case-sensitive. See 'DigitalExperienceBundle' in the Metadata API Developer Guide for more information. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_digitalexperiencebundle.htm)
         /// 
-        /// When creating a site with the Build Your Own (LWR) template, you must also specify the AuthenticationType value using the format templateParams.AuthenticationType=value, where value is AUTHENTICATED, UNAUTHENTICATED, or AUTHENTICATED_WITH_PUBLIC_ACCESS. Name and values are case-sensitive. See "ExperienceBundle" in the Metadata API Developer Guide for more information.
+        /// The site creation process is an async job that generates a jobId. To check the site creation status, query the BackgroundOperation object and enter the jobId as the Id. See ‘BackgroundOperation’ in the Object Reference for the Salesforce Platform for more information. (https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_backgroundoperation.htm)
         /// 
-        /// When you execute this command, it creates the site in preview status, which means that it isn’t yet live. After you finish building your site, you can make it live.
+        /// If the job doesn’t complete within 10 minutes, it times out. You receive an error message and must restart the site creation process. Completed jobs expire after 24 hours and are removed from the database.
         /// 
-        /// If you have an Experience Builder site, publish the site using the sfdx force:community:publish command to make it live.
+        /// When you run this command, it creates the site in preview status, which means that the site isn't yet live. After you finish building your site, you can make it live.
         /// 
-        /// If you have a Salesforce Tabs + Visualforce site, activate the site to make it live by updating the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
+        /// If you have an Experience Builder site, publish the site using the "community publish" command to make it live.
         /// 
-        /// For Experience Builder sites, activating the site just sends out a welcome email to site members.
+        /// If you have a Salesforce Tabs + Visualforce site, to activate the site and make it live, update the status field of the Network type in Metadata API. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_network.htm) Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
         /// 
-        /// Examples:
-        ///    $ sfdx force:community:create --name "My Customer Site" --templatename "Customer Service" --urlpathprefix customers --description "My customer site"
-        ///    $ sfdx force:community:create -n partnercentral -t "Partner Central" -p partners
-        ///    $ sfdx force:community:create -n lwrsite -t "Build Your Own (LWR)" -p lwrsite templateParams.AuthenticatedType=UNAUTHENTICATED
+        /// For Experience Builder sites, activating the site sends a welcome email to site members.
         /// </remarks>
         /// <example>
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
         /// 
-        /// When creating a site with the Build Your Own (LWR) template, you must also specify the AuthenticationType value using the format templateParams.AuthenticationType=value, where value is AUTHENTICATED, UNAUTHENTICATED, or AUTHENTICATED_WITH_PUBLIC_ACCESS. Name and values are case-sensitive. See "ExperienceBundle" in the Metadata API Developer Guide for more information.
-        /// 
-        /// When you execute this command, it creates the site in preview status, which means that it isn’t yet live. After you finish building your site, you can make it live.
-        /// 
-        /// If you have an Experience Builder site, publish the site using the sfdx force:community:publish command to make it live.
-        /// 
-        /// If you have a Salesforce Tabs + Visualforce site, activate the site to make it live by updating the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
-        /// 
-        /// For Experience Builder sites, activating the site just sends out a welcome email to site members.
-        /// 
-        /// Examples:
-        ///    $ sfdx force:community:create --name "My Customer Site" --templatename "Customer Service" --urlpathprefix customers --description "My customer site"
-        ///    $ sfdx force:community:create -n partnercentral -t "Partner Central" -p partners
-        ///    $ sfdx force:community:create -n lwrsite -t "Build Your Own (LWR)" -p lwrsite templateParams.AuthenticatedType=UNAUTHENTICATED
-        /// force:community:create [name=value...] -n <string> -t <string> -p <string> [-d <string>] [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]
+        /// Create an Experience Cloud site using template 'Customer Service' and URL path prefix 'customers':
+        /// <%= config.bin %> <%= command.id %> --name 'My Customer Site' --template-name 'Customer Service' --url-path-prefix customers --description 'My customer site'
+        /// Create a site using 'Partner Central' template:
+        /// <%= config.bin %> <%= command.id %> --name partnercentral --template-name 'Partner Central' --url-path-prefix partners
+        /// Create a site using the 'Build Your Own (LWR)' template with authentication type of UNAUTHENTICATED:
+        /// <%= config.bin %> <%= command.id %> --name lwrsite --template-name 'Build Your Own (LWR)' --url-path-prefix lwrsite templateParams.AuthenticationType=UNAUTHENTICATED
         /// </example>
         [Command("create")]
-        public SFDXResponse Create(CommunityCreateOptions p_Options)
+        public SfdxResponse Create(CommunityCreateOptions options)
         {
-            return ExecuteCommand<CommunityCreateOptions>(nameof(Create), p_Options);
+            return ExecuteCommand(nameof(Create), options);
         }
 
         /// <summary>
-        /// publishes an Experience Builder site to make it live
-        /// 
-        /// When you publish an Experience Builder site for the first time, you make the site’s URL live and enable login access for site members.
-        /// 
-        /// Additionally, to send a welcome email to all site members, you must activate the site. (Activation is also required to successfully set up SEO for Experience Builder sites.) To activate a site, update the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
-        /// 
-        /// Subsequently, each time you publish the site, you update the live site with all changes made to the site since it was last published.
-        /// 
-        /// An email notification informs you when your changes are live.
-        /// 
-        /// Example:
-        ///    $ sfdx force:community:publish --name "My Customer Site"
+        /// See 'Which Experience Cloud Template Should I Use?' (https://help.salesforce.com/s/articleView?id=sf.siteforce_commtemp_intro.htm&type=5) in Salesforce Help for more information about the different template types available for Experience Cloud.
         /// </summary>
         /// <remarks>
-        /// publishes an Experience Builder site to make it live
-        /// 
-        /// When you publish an Experience Builder site for the first time, you make the site’s URL live and enable login access for site members.
-        /// 
-        /// Additionally, to send a welcome email to all site members, you must activate the site. (Activation is also required to successfully set up SEO for Experience Builder sites.) To activate a site, update the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
-        /// 
-        /// Subsequently, each time you publish the site, you update the live site with all changes made to the site since it was last published.
-        /// 
-        /// An email notification informs you when your changes are live.
-        /// 
-        /// Example:
-        ///    $ sfdx force:community:publish --name "My Customer Site"
+        /// See 'Which Experience Cloud Template Should I Use?' (https://help.salesforce.com/s/articleView?id=sf.siteforce_commtemp_intro.htm&type=5) in Salesforce Help for more information about the different template types available for Experience Cloud.
         /// </remarks>
         /// <example>
-        /// When you publish an Experience Builder site for the first time, you make the site’s URL live and enable login access for site members.
         /// 
-        /// Additionally, to send a welcome email to all site members, you must activate the site. (Activation is also required to successfully set up SEO for Experience Builder sites.) To activate a site, update the status field of the Network type in the Metadata API. Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
+        /// Retrieve the template list from an org with alias my-scratch-org:
+        /// <%= config.bin %> <%= command.id %> --target-org my-scratch-org
+        /// </example>
+        [Command("list template")]
+        public SfdxResponse ListTemplate(CommunityListTemplateOptions options)
+        {
+            return ExecuteCommand(nameof(ListTemplate), options);
+        }
+
+        /// <summary>
+        /// Each time you publish a site, you update the live site with the most recent updates. When you publish an Experience Builder site for the first time, you make the site's URL live and enable login access for site members.
         /// 
-        /// Subsequently, each time you publish the site, you update the live site with all changes made to the site since it was last published.
+        /// In addition to publishing, you must activate a site to send a welcome email to all site members. Activation is also required to set up SEO for Experience Builder sites. To activate a site, update the status field of the Network type in Metadata API. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_network.htm) Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
         /// 
-        /// An email notification informs you when your changes are live.
+        /// An email notification informs you when your changes are live on the published site. The site publish process is an async job that generates a jobId. To check the site publish status manually, query the BackgroundOperation object and enter the jobId as the Id. See ‘BackgroundOperation’ in the Object Reference for the Salesforce Platform for more information. (https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_backgroundoperation.htm)
         /// 
-        /// Example:
-        ///    $ sfdx force:community:publish --name "My Customer Site"
-        /// force:community:publish -n <string> [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]
+        /// If the job doesn’t complete within 15 minutes, it times out. You receive an error message and must restart the site publish process. Completed jobs expire after 24 hours and are removed from the database.
+        /// </summary>
+        /// <remarks>
+        /// Each time you publish a site, you update the live site with the most recent updates. When you publish an Experience Builder site for the first time, you make the site's URL live and enable login access for site members.
+        /// 
+        /// In addition to publishing, you must activate a site to send a welcome email to all site members. Activation is also required to set up SEO for Experience Builder sites. To activate a site, update the status field of the Network type in Metadata API. (https://developer.salesforce.com/docs/atlas.en-us.api_meta.meta/api_meta/meta_network.htm) Alternatively, in Experience Workspaces, go to Administration | Settings, and click Activate.
+        /// 
+        /// An email notification informs you when your changes are live on the published site. The site publish process is an async job that generates a jobId. To check the site publish status manually, query the BackgroundOperation object and enter the jobId as the Id. See ‘BackgroundOperation’ in the Object Reference for the Salesforce Platform for more information. (https://developer.salesforce.com/docs/atlas.en-us.object_reference.meta/object_reference/sforce_api_objects_backgroundoperation.htm)
+        /// 
+        /// If the job doesn’t complete within 15 minutes, it times out. You receive an error message and must restart the site publish process. Completed jobs expire after 24 hours and are removed from the database.
+        /// </remarks>
+        /// <example>
+        /// 
+        /// Publish the Experience Builder site with name "My Customer Site':
+        /// <%= config.bin %> <%= command.id %> --name 'My Customer Site'
         /// </example>
         [Command("publish")]
-        public SFDXResponse Publish(CommunityPublishOptions p_Options)
+        public SfdxResponse Publish(CommunityPublishOptions options)
         {
-            return ExecuteCommand<CommunityPublishOptions>(nameof(Publish), p_Options);
-        }
-
-        /// <summary>
-        /// retrieves the list of templates available in your org
-        /// 
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
-        /// 
-        /// Example:
-        ///    $ sfdx force:community:template:list
-        /// </summary>
-        /// <remarks>
-        /// retrieves the list of templates available in your org
-        /// 
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
-        /// 
-        /// Example:
-        ///    $ sfdx force:community:template:list
-        /// </remarks>
-        /// <example>
-        /// See "Which Experience Cloud Template Should I Use?" in Salesforce Help for more information about the different template types available for Experience Cloud.
-        /// 
-        /// Example:
-        ///    $ sfdx force:community:template:list
-        /// force:community:template:list [-u <string>] [--apiversion <string>] [--json] [--loglevel trace|debug|info|warn|error|fatal|TRACE|DEBUG|INFO|WARN|ERROR|FATAL]
-        /// </example>
-        [Command("template:list")]
-        public SFDXResponse TemplateList(CommunityTemplateListOptions p_Options)
-        {
-            return ExecuteCommand<CommunityTemplateListOptions>(nameof(TemplateList), p_Options);
+            return ExecuteCommand(nameof(Publish), options);
         }
     }
 }
