@@ -1,6 +1,4 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using System;
+﻿using System;
 using System.Diagnostics;
 using System.Reflection;
 
@@ -9,31 +7,31 @@ namespace sfdx4csharpClient.Core.Attributes
     [AttributeUsage(AttributeTargets.Property)]
     public class SwitchNameAttribute : Attribute
     {
-        private readonly string m_Value;
+        private readonly string _value;
 
         /// <summary>
         /// Constructor.
         /// </summary>
-        /// <param name="p_Value">SFDX Switch parameter.</param>
+        /// <param name="value">SFDX Switch parameter.</param>
         /// <example>For the JSON option, the switch parameter should be "--json".</example>
-        public SwitchNameAttribute(string p_Value)
+        public SwitchNameAttribute(string value)
         {
             // Switch name can be empty (ex: OrgCreateOptions.Expression), but not null.
-            Debug.Assert(p_Value != null);
+            Debug.Assert(value != null);
 
-            m_Value = p_Value;
+            _value = value;
         }
 
         /// <summary>
         /// Gets the attribute value for the given property.
         /// </summary>
-        /// <param name="p_PropertyInfo">Property with a SwitchNameAttribute to get.</param>
+        /// <param name="propertyInfo">Property with a SwitchNameAttribute to get.</param>
         /// <returns>The attribute value.</returns>
-        public static string GetSwitchNameValue(PropertyInfo p_PropertyInfo)
+        public static string GetSwitchNameValue(PropertyInfo propertyInfo)
         {
-            Debug.Assert(p_PropertyInfo != null);
+            Debug.Assert(propertyInfo != null);
 
-            return p_PropertyInfo.GetCustomAttribute<SwitchNameAttribute>()?.m_Value;
+            return propertyInfo.GetCustomAttribute<SwitchNameAttribute>()?._value;
         }
     }
 }
