@@ -1,6 +1,4 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using System.Diagnostics;
+﻿using System.Diagnostics;
 
 namespace sfdx4csharpClient.Core
 {
@@ -9,32 +7,32 @@ namespace sfdx4csharpClient.Core
     /// </summary>
     public class CommandExecutioner
     {
-        private readonly CommandRunner m_CommandRunner;
+        private readonly CommandRunner _commandRunner;
 
         /// <summary>
         /// Constructor.
         /// </summary>
-        /// <param name="p_CommandRunner">The command executor that will run the commands.</param>
-        public CommandExecutioner(CommandRunner p_CommandRunner)
+        /// <param name="commandRunner">The command executor that will run the commands.</param>
+        public CommandExecutioner(CommandRunner commandRunner)
         {
-            Debug.Assert(p_CommandRunner != null);
+            Debug.Assert(commandRunner != null);
 
-            m_CommandRunner = p_CommandRunner;
+            _commandRunner = commandRunner;
         }
 
         /// <summary>
         /// Execute a command's method and return the serialized result.
         /// </summary>
         /// <typeparam name="T">Type of the command's method options.</typeparam>
-        /// <param name="p_Command">Current command's method.</param>
-        /// <param name="p_RequestOptions">Current command's method options.</param>
-        /// <returns>SFDX output informations.</returns>
-        public SFDXOutput Execute<T>(string p_Command, T p_RequestOptions) where T : SFDXOptions
+        /// <param name="command">Current command's method.</param>
+        /// <param name="requestOptions">Current command's method options.</param>
+        /// <returns>SFDX output information.</returns>
+        public SfdxOutput Execute<T>(string command, T requestOptions) where T : SfdxOptions
         {
-            Debug.Assert(p_Command != null);
+            Debug.Assert(command != null);
 
-            string completeCommand = CommandBuilder<T>.Build(p_Command, p_RequestOptions);
-            return m_CommandRunner.RunCommand(completeCommand);
+            var completeCommand = CommandBuilder<T>.Build(command, requestOptions);
+            return _commandRunner.RunCommand(completeCommand);
         }
     }
 }
