@@ -1,6 +1,4 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using System.Reflection;
 using sfdx4csharpClient.Core.Attributes;
 
@@ -9,32 +7,32 @@ namespace sfdx4csharpClient.Core
     /// <summary>
     /// SFDX CLI command's wrapper.
     /// </summary>
-    public abstract class SFDXCommand
+    public abstract class SfdxCommand
     {
-        private readonly CommandExecutioner m_CommandExecutioner;
+        private readonly CommandExecutioner _commandExecutioner;
 
-        protected SFDXCommand(CommandExecutioner p_CommandExecutioner)
+        protected SfdxCommand(CommandExecutioner commandExecutioner)
         {
-            Debug.Assert(p_CommandExecutioner != null);
+            Debug.Assert(commandExecutioner != null);
 
-            m_CommandExecutioner = p_CommandExecutioner;
+            _commandExecutioner = commandExecutioner;
         }
 
-        protected SFDXResponse ExecuteCommand<TOptions>(string p_MethodName,
-            TOptions p_Options) where TOptions : SFDXOptions, new()
+        protected SfdxResponse ExecuteCommand<TOptions>(string methodName, TOptions commandOptions) where TOptions : SfdxOptions, new()
         {
-            Debug.Assert(p_MethodName != null);
+            Debug.Assert(methodName != null);
 
-            MethodInfo methodInfo = GetType().GetMethod(p_MethodName);
+            var methodInfo = GetType().GetMethod(methodName);
             Debug.Assert(methodInfo != null);
 
-            string apiCommandClass = CommandClassAttribute.GetCommandClassValue(GetType());
-            string apiCommand = CommandAttribute.GetCommandValue(methodInfo);
-            string command = $"{apiCommandClass}:{apiCommand}";
+            var apiCommandClass = CommandClassAttribute.GetCommandClassValue(GetType());
+            var apiCommand = CommandAttribute.GetCommandValue(methodInfo);
+            var command = $"{apiCommandClass} {apiCommand}";
 
-            SFDXOptions options = p_Options ?? new TOptions();
-            SFDXOutput output = m_CommandExecutioner.Execute(command, options);
-            return new SFDXResponse {
+            SfdxOptions options = commandOptions ?? new TOptions();
+            var output = _commandExecutioner.Execute(command, options);
+            return new SfdxResponse
+            {
                 AdditionalInfo = output,
                 Result = options.json ? ResponseParser.Parse(output.RawOutput) : null
             };
