@@ -1,13 +1,11 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using Newtonsoft.Json.Linq;
+﻿using Newtonsoft.Json.Linq;
 
 namespace sfdx4csharpClient.Core
 {
     /// <summary>
     /// SFDX CLI Json Response's wrapper
     /// </summary>
-    public class SFDXJsonResponse
+    public class SfdxJsonResponse
     {
         /// <summary>
         /// Deserialized from cli output. Represents the parsed output of the executed command.
