@@ -1,11 +1,9 @@
-// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
 namespace sfdx4csharpClient.Core
 {
     /// <summary>
     /// SFDX CLI Output's wrapper
     /// </summary>
-    public class SFDXOutput
+    public class SfdxOutput
     {
         /// <summary>
         /// SFDX cli command used with parameters and flags.
