@@ -1,13 +1,11 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using Newtonsoft.Json.Linq;
+﻿using Newtonsoft.Json.Linq;
 
 namespace sfdx4csharpClient.Core
 {
     /// <summary>
     /// SFDX CLI Response's wrapper
     /// </summary>
-    public class SFDXResponse
+    public class SfdxResponse
     {
         /// <summary>
         /// Deserialized from cli output. Represents the parsed output of the executed command.
@@ -17,6 +15,6 @@ namespace sfdx4csharpClient.Core
         /// <summary>
         /// Deserialized from cli output. Represents the status code from the executed command.
         /// </summary>
-        public SFDXOutput AdditionalInfo { get; set; }
+        public SfdxOutput AdditionalInfo { get; set; }
     }
 }
