@@ -1,60 +1,60 @@
-﻿// Copyright (c) 2005-2020, Coveo Solutions Inc.
-
-using System;
+﻿using System;
 using System.Diagnostics;
 using System.IO;
+using Newtonsoft.Json;
 using sfdx4csharpClient;
 using sfdx4csharpClient.Core;
 
 namespace sfdx4csharp
 {
-    class Program
+    public class Program
     {
-        const string SFDX_PATH_ENV_KEY = "SFDX_PATH";
+        private const string SFDX_PATH_ENV_KEY = "SF_PATH";
 
-        static void Main(string[] args)
+        static void Main()
         {
             Console.WriteLine("Verifying sfdx version.");
             var cliPath = Environment.GetEnvironmentVariable(SFDX_PATH_ENV_KEY);
             Debug.Assert(cliPath != null);
 
-            CommandRunner runner = new CommandRunner(cliPath);
-            string version = runner.RunCommand("--version").RawOutput;
+            var runner = new CommandRunner(cliPath);
+            var version = runner.RunCommand("--version").RawOutput;
             Console.WriteLine($"Using SFDX version: {version}");
 
-            Console.WriteLine("Simple debug test using api:display command.");
-            var client = new SFDXClient(cliPath);
-
-            string projectFolder = "C:\\TmpSfdxProject";
+            var client = new SfdxClient(cliPath);
+            var projectFolder = "C:\\TmpSfdxProject";
             if (Directory.Exists(projectFolder)) {
                 Directory.Delete(projectFolder, true);
             }
 
             Directory.CreateDirectory(projectFolder);
             client.ChangeWorkingFolder(projectFolder);
-            client.Project.Create(new ProjectCreateOptions { projectname = "Test" });
+            Console.WriteLine("Listing the existing orgs.");
+            client.Org.List(new OrgListOptions { All = true});
 
-            client.ChangeWorkingFolder(Path.Combine(projectFolder, "Test"));
             Console.WriteLine("Creating test project.");
-            Console.WriteLine(client.Project.Create(new ProjectCreateOptions {
-                projectname = "test"
-            }));
+            client.Project.Generate(new ProjectGenerateOptions { Name = "Test" });
+            client.ChangeWorkingFolder(Path.Combine(projectFolder, "Test"));
 
             Console.WriteLine("Creating a scratch org.");
-            Console.WriteLine(client.Org.Create(new OrgCreateOptions {
-                definitionfile = "config\\project-scratch-def.json",
-                setalias = "Toto",
-                wait = 5
-            }));
+            client.Org.CreateScratch(new OrgCreateScratchOptions
+            {
+                DefinitionFile = "config\\project-scratch-def.json",
+                Alias = "Toto",
+                Wait = "5"
+            });
 
             Console.WriteLine("Project status:");
-            Console.WriteLine(client.Source.Status(new SourceStatusOptions {
-                targetusername = "Toto"
-            }));
+            var response = client.Project.RetrievePreview(new ProjectRetrievePreviewOptions
+            {
+                TargetOrg = "Toto"
+            });
 
-            client.Org.Delete(new OrgDeleteOptions {
-                noprompt = true,
-                targetusername = "Toto"
+            Console.WriteLine(JsonConvert.SerializeObject(response));
+
+            client.Org.DeleteScratch(new OrgDeleteScratchOptions {
+                NoPrompt = true,
+                TargetOrg = "Toto"
             });
 
             Console.WriteLine("Dummy run done!");
